Refactor joint inversion drivers and move more core methods to the base
