+++ simpeg_drivers/joint/joint_cross_gradient/driver.py
@@ -44,157 +44,13 @@ class JointCrossGradientDriver(BaseJointDriver):
         with fetch_active_workspace(self.workspace, mode="r+"):
             self.initialize()
 
-    def validate_create_models(self):
-        """Create stacked model vectors from all drivers provided."""
-        for model_type in self.models.model_types:
-            model = np.zeros(self.models.n_active * len(self.mapping))
-
-            for child_driver in self.drivers:
-                model_local_values = getattr(child_driver.models, model_type)
-
-                if model_local_values is not None:
-                    projection = child_driver.data_misfit.model_map.deriv(model).T
-
-                    if isinstance(model_local_values, float):
-                        model_local_values = (
-                            np.ones(projection.shape[1]) * model_local_values
-                        )
-
-                    norm = np.array(np.sum(projection, axis=1)).flatten()
-                    model += (projection * model_local_values) / (norm + 1e-8)
-
-            if model is not None:
-                getattr(self.models, f"_{model_type}").model = model
-
-    @property
-    def directives(self):
-        if getattr(self, "_directives", None) is None and not self.params.forward_only:
-            with fetch_active_workspace(self.workspace, mode="r+"):
-                self._directives = DirectivesFactory(self)
-
-                directives_list = []
-                count = 0
-                for driver in self.drivers:
-                    driver_directives = driver.directives
-                    n_tiles = len(driver.data_misfit.objfcts)
-
-                    for name in [
-                        "save_iteration_data_directive",
-                        "save_iteration_residual_directive",
-                        "save_iteration_apparent_resistivity_directive",
-                    ]:
-                        directive = getattr(driver_directives, name)
-                        if directive is not None:
-                            directive.joint_index = [
-                                count + ii for ii in range(n_tiles)
-                            ]
-                            directives_list.append(directive)
-
-                    save_model = driver_directives.save_iteration_model_directive
-                    save_model.label = driver.params.physical_property
-                    save_model.transforms = [
-                        driver.data_misfit.model_map,
-                        *save_model.transforms,
-                    ]
-
-                    directives_list.append(save_model)
-                    directives_list.append(
-                        directives.SaveLPModelGroup(
-                            driver.inversion_mesh.entity,
-                            self._directives.update_irls_directive,
-                        )
-                    )
-
-                    if driver_directives.vector_inversion_directive is not None:
-                        directives_list.append(
-                            driver_directives.vector_inversion_directive
-                        )
-
-                    if driver_directives.save_property_group is not None:
-                        directives_list.append(driver_directives.save_property_group)
-
-                    save_sensitivities = driver_directives.save_sensitivities_directive
-                    if save_sensitivities is not None:
-                        save_sensitivities.transforms = [
-                            driver.data_misfit.model_map,
-                            *save_sensitivities.transforms,
-                        ]
-                        directives_list.append(save_sensitivities)
-
-                    count += n_tiles
-
-                for driver, wire in zip(self.drivers, self.wires, strict=True):
-                    factory = SaveModelGeoh5Factory(driver.params)
-                    factory.factory_type = driver.params.inversion_type
-                    model_directive = factory.build(
-                        inversion_object=self.inversion_mesh,
-                        active_cells=self.models.active_cells,
-                        name="Model",
-                    )
-
-                    model_directive.label = driver.params.physical_property
-                    if (
-                        getattr(driver.params, "model_type", None)
-                        == "Resistivity (Ohm-m)"
-                    ):
-                        model_directive.label = "resistivity"
-
-                    model_directive.transforms = [wire, *model_directive.transforms]
-                    directives_list.append(model_directive)
-
-                    if driver.directives.save_property_group is not None:
-                        directives_list.append(
-                            directives.SavePropertyGroup(
-                                self.inversion_mesh.entity,
-                                group_type=GroupTypeEnum.DIPDIR,
-                                channels=["declination", "inclination"],
-                            )
-                        )
-
-                    directives_list.append(
-                        directives.SaveLPModelGroup(
-                            self.inversion_mesh.entity,
-                            self._directives.update_irls_directive,
-                        )
-                    )
-
-                directives_list.append(self._directives.save_iteration_log_files)
-                self._directives.directive_list = (
-                    self._directives.inversion_directives + directives_list
-                )
-        return self._directives
-
     def get_regularization(self):
         """
         Create a flat ComboObjectiveFunction from all drivers provided and
         add cross-gradient regularization for all combinations of model parameters.
         """
         regularizations = super().get_regularization()
-        reg_list = regularizations.objfcts
-        multipliers = regularizations.multipliers
-        reg_dict = {reg.mapping: reg for reg in reg_list}
-        for driver in self.drivers:
-            reg_block = []
-            for mapping in driver.mapping:
-                reg_block.append(reg_dict[self._mapping[driver, mapping]])
-
-            # Pass down regularization parameters from driver.
-            for param in [
-                "alpha_s",
-                "length_scale_x",
-                "length_scale_y",
-                "length_scale_z",
-                "s_norm",
-                "x_norm",
-                "y_norm",
-                "z_norm",
-                "gradient_type",
-            ]:
-                if getattr(self.params, param) is None:
-                    for reg in reg_block:
-                        setattr(reg, param, getattr(driver.params, param))
-
-            driver.regularization = ComboObjectiveFunction(objfcts=reg_block)
+        reg_list, multipliers = self._overload_regularization(regularizations)
 
         for label, driver_pairs in zip(
             ["a_b", "c_a", "c_b"], combinations(self.drivers, 2), strict=False
@@ -222,52 +78,3 @@ class JointCrossGradientDriver(BaseJointDriver):
                     )
 
         return ComboObjectiveFunction(objfcts=reg_list, multipliers=multipliers)
-
-    @property
-    def wires(self):
-        """Model projections"""
-        if self._wires is None:
-            collection = []
-            start = 0
-            for n_values in self.n_values:
-                collection.append(
-                    maps.Projection(
-                        int(np.sum(self.n_values)), slice(start, start + n_values)
-                    )
-                )
-                start += n_values
-
-            self._wires = collection
-
-        return self._wires
-
-    @property
-    def mapping(self):
-        """Create a list of all mappings augmented with projection to global problem"""
-        if self._mapping is None:
-            mappings = {}
-            start = 0
-            n_values = int(np.sum(self.models.active_cells))
-            for driver in self.drivers:
-                for mapping in driver.mapping:
-                    mappings[driver, mapping] = maps.Projection(
-                        int(np.sum(self.n_values)), slice(start, start + n_values)
-                    )
-                    start += n_values
-
-            self._mapping = mappings
-
-        return self._mapping.values()
-
-    @property
-    def n_values(self):
-        """Number of values in the model"""
-        if self._n_values is None:
-            n_values = self.models.n_active
-            count = []
-            for driver in self.drivers:
-                n_comp = 3 if driver.inversion_data.vector else 1
-                count.append(n_values * n_comp)
-            self._n_values = count
-
-        return self._n_values
